backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
sqlalchemy>=2.0.0
aiosqlite>=0.20.0
anthropic>=0.40.0
python-docx>=1.1.0
pydantic>=2.9.0
pydantic-settings>=2.6.0
python-multipart>=0.0.12
jinja2>=3.1.0
httpx>=0.27.0
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
